app/requirements.txt
Flask
mysql-connector
pandas
sqlalchemy
mysqlclient
Pillow
keras
tensorflow
numpy
flake8
pytest
black